backend/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-spa==0.3.6
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
gunicorn==23.0.0
ipython==8.28.0
ipython-genutils==0.2.0
jedi==0.19.1
matplotlib-inline==0.1.7
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
whitenoise==5.3.0
